renaming package and adding json to object mapper
