+++ src/pactVerificationTest/resources/judges.json
@@ -1,0 +1,8 @@
+{
+  "judiciary": {
+    "johTitle": "His Honour",
+    "johNameSurname": "John Smith",
+    "role": "judge",
+    "johKnownAs": "His Honour Judge Smith"
+  }
+}
